Update scenery files - Version 0.2.0

+++ R/lods.R
@@ -69,14 +69,14 @@ checkAllLods <- function(modelLibDir) {
 
 #' Fix Lods
 #'
-#' @param PackageSourcesDir
-#' @param invalids
+#' @param PackageSourcesDir Path to PackageSources directory.
+#' @param invalids List with invalid files from `checkAllLods`.
+#' @param deleteTextures Whether to delete textures in modelLib/texture related to the corrupted data.
 #'
-#' @return
 #' @export
 #'
 #' @examples
-fixLods <- function(PackageSourcesDir, invalids) {
+fixLods <- function(PackageSourcesDir, invalids, deleteTextures = TRUE) {
 
   # Delete invalid XML from modelLib
   message("Removendo arquivos XML inválidos em modelLib")
@@ -86,14 +86,16 @@ fixLods <- function(PackageSourcesDir, invalids) {
   message("----------------------------")
 
   # Delete correspondent textures
-  ids <- stringr::str_remove(sapply(invalids, "[[", 3), ".xml$")
-  texToRemove <- list.files(file.path(PackageSourcesDir, "modelLib", "texture"),
-                            paste0(ids, collapse = "|"),
-                            all.files = TRUE,
-                                      full.names = TRUE)
-  statusPNG <- file.remove(texToRemove)
-  message(sum(statusPNG), " arquivos PNG (texture) removidos")
-  message("----------------------------")
+  if (isTRUE(deleteTextures)) {
+    ids <- stringr::str_remove(sapply(invalids, "[[", 3), ".xml$")
+    texToRemove <- list.files(file.path(PackageSourcesDir, "modelLib", "texture"),
+                              paste0(ids, collapse = "|"),
+                              all.files = TRUE,
+                              full.names = TRUE)
+    statusPNG <- file.remove(texToRemove)
+    message(sum(statusPNG), " arquivos PNG (texture) removidos")
+    message("----------------------------")
+  }
 
   # Clean corrupted guids from objects.xml
   invalidGuids <- sapply(invalids, "[[", 2)
